add: trojan_sss to inject to explorer.exe
The trojan works as follows:
- Create a small space to store the dll path
- Create a thread to run LoadLibraryA(dll_path)

The old sss trojan makes the code stuck on while loop, the DllMain should run less than 300ms. We create a thread to run the while loop.

The trojan is still a command line application, will change to WinMain so that user don't know that it was run.